+++ CHANGELOG.rst
@@ -3,7 +3,7 @@ Next (tbd)
 
 * `#15 <https://github.com/pytest-dev/pytest-reportlog/issues/15>`_: switch recommended file extension to `jsonl <https://jsonlines.org/>`__.
 * `#17 <https://github.com/pytest-dev/pytest-reportlog/issues/17>`_: add compression support for ``.bz2``/``.gz``/``.xz`` log file extensions.
-* `#39 <https://github.com/pytest-dev/pytest-reportlog/issues/39>`_: allow not capturing logs for passed tests
+* `#39 <https://github.com/pytest-dev/pytest-reportlog/issues/39>`_: new ``--report-log-exclude-logs-on-passed-tests`` option allows not capturing logs of passing tests.
 
 
 0.3.0 (2023-04-26)